feat: support yaml file for Hazelcast configuration
Supporting Yaml configuration file should ease the overriding config in Helm chart.
Currently, with the XML file, we can't use Helm/Go template function to merge properties.

With the YAML, we should be able to define a default configuration in the Helm chart
and, let users overrides part of that configuration with values, using https://rm3l.org/merging-dynamic-config-data-in-helm-charts/